Put the snappyimage files in a portal-friendly subdirectory. Added a photo.
