+++ README.md
@@ -1,0 +1,39 @@
+# Snappybot
+
+This is a directy copy off of Chris Dail's repo. I have to create a temporary rep in mine in order to see the README.md rendered. When I get all this together I'll merge it back into the clone of his and request a merge to the real ... then delete this repo.
+
+## Links
+
+Arduino Compatible DIY Motor Smart Robot Car Chassis Kit 4 WD<br>
+http://www.gearbest.com/development-boards/pp_157852.html
+
+Adafruit Motor Shield<br>
+https://www.adafruit.com/products/1438 
+
+PCA9685: 16-channel, 12-bit PWM Fm+ I2C-bus LED controller (on the Adafruit Motor Shield)<br>
+http://www.nxp.com/products/power-management/lighting-driver-and-controller-ics/i2c-led-display-control/16-channel-12-bit-pwm-fm-plus-ic-bus-led-controller:PCA9685
+
+SparkFun RGB and Gesture Sensor APDS-9960<br>
+https://www.sparkfun.com/products/12787
+
+## Hardware
+
+![](https://github.com/lotsandlotsofrobots/snappybot/blob/master/art/profile.jpg)
+
+Right head light (sitting in car) is S8. The left is S10. 
+Left front M4, right front M3, left front m1, left back m2.
+
+## Software
+
+motor_controller_startup( 0, 3,  1, 2,  False)
+
+06.A4.D9 ch:6 Net:0xABCD
+
+
+
+
+
+
+
+
+
