+++ portal/snappyImages/.gitignore
@@ -1,0 +1,4 @@
+/ada_moto_v2.pyo
+/apds_9960_rgb_gesture_sensor.pyo
+/PCA9685.pyo
+/pyduinoincludes.pyo
