+++ portal/snappyImages/ada_moto_v2.py
